code review issues
